word-addon/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acb-large-print"
version = "1.0.0"
description = "Audit, fix, and template Word documents for ACB Large Print Guidelines compliance"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jeff Bishop", email = "jeff@bitsonline.org"},
]
keywords = ["accessibility", "large-print", "ACB", "WCAG", "Word", "docx"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: General",
    "Topic :: Adaptive Technologies",
]
dependencies = [
    "python-docx>=1.1.0",
    "mammoth>=1.8.0",
    "wxPython>=4.2.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0",
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
acb-large-print = "acb_large_print.__main__:main"

[project.gui-scripts]
acb-large-print-gui = "acb_large_print.gui:launch_gui"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
